move population script
